+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 # CHANGELOG
 
-## [2.0.7-dev.0] - 2021-11-04
+## [2.0.7-dev.2] - 2021-11-04
+### Added
 - Add `options.ignoreJsonIncompatibleTypes`
+### Changed
+- Default option of `strictNullChecks` changed to `false`
+### Fixed
+- `null` and `undefined` is convertable when `prune`
 
 ## [2.0.6] - 2021-11-04
 - Update dependencies